requirements.txt
openai
pinecone
pandas
tiktoken
python-dotenv
numpy
tqdm
langchain-openai
fastapi
uvicorn[standard]
